+++ datatypes/src/jvmTest/kotlin/PublicKeyTest.kt
@@ -23,7 +23,7 @@ import java.security.interfaces.RSAPublicKey
 class PublicKeyTest : FreeSpec({
     "EC" - {
         withData(256, 384, 521) { bits ->
-            val keys = List<ECPublicKey>(100) {
+            val keys = List<ECPublicKey>(256000 / bits) {
                 val ecKp = KeyPairGenerator.getInstance("EC").apply {
                     initialize(bits)
                 }.genKeyPair()
@@ -38,38 +38,32 @@ class PublicKeyTest : FreeSpec({
                 keys
             ) { pubKey ->
 
-                val own: CryptoPublicKey = CryptoPublicKey.Ec.fromJcaKey(pubKey).getOrThrow() as CryptoPublicKey.Ec
-                own.shouldNotBeNull()
-
-                val pubKey2: CryptoPublicKey = CryptoPublicKey.Ec.fromJcaKey(pubKey).getOrThrow()
-                val iosDerived: CryptoPublicKey = CryptoPublicKey.fromIosEncoded(own.iosEncoded)
-                val keyIdDerived: CryptoPublicKey = CryptoPublicKey.fromKeyId(own.keyId)
-
-                pubKey2.hashCode() shouldBe own.hashCode()
-                pubKey2 shouldBe own
-                iosDerived.hashCode() shouldBe own.hashCode()
-                iosDerived shouldBe own
-                keyIdDerived.hashCode() shouldBe own.hashCode()
-                keyIdDerived shouldBe own
-
+                val own = CryptoPublicKey.Ec.fromJcaKey(pubKey).getOrThrow()
                 println(Json.encodeToString(own))
                 println(own.iosEncoded.encodeToString(Base16()))
                 println(own.encodeToDer().encodeToString(Base16()))
                 println(own.keyId)
                 own.encodeToDer() shouldBe pubKey.encoded
-
+                CryptoPublicKey.fromKeyId(own.keyId) shouldBe own
                 own.getPublicKey().encoded shouldBe pubKey.encoded
                 CryptoPublicKey.decodeFromTlv(Asn1Element.parse(own.encodeToDer()) as Asn1Sequence) shouldBe own
-
-
             }
+        }
+
+        "Equality tests" {
+            val keyPair = KeyPairGenerator.getInstance("EC").also { it.initialize(256) }.genKeyPair()
+            val pubKey1 = CryptoPublicKey.decodeFromDer(keyPair.public.encoded)
+            val pubKey2 = CryptoPublicKey.decodeFromDer(keyPair.public.encoded)
+
+            pubKey1.hashCode() shouldBe pubKey2.hashCode()
+            pubKey1 shouldBe pubKey2
         }
 
     }
 
     "RSA" - {
         withData(512, 1024, 2048, 3072, 4096) { bits ->
-            val keys = List<RSAPublicKey>(100) {
+            val keys = List<RSAPublicKey>(13000 / bits) {
                 val rsaKP = KeyPairGenerator.getInstance("RSA").apply {
                     initialize(bits)
                 }.genKeyPair()
@@ -84,25 +78,15 @@ class PublicKeyTest : FreeSpec({
                 keys
             ) { pubKey ->
 
-                val own: CryptoPublicKey.Rsa = CryptoPublicKey.fromJcaKey(pubKey).getOrThrow() as CryptoPublicKey.Rsa
-                own.shouldNotBeNull()
-                val pubKey2: CryptoPublicKey.Rsa = CryptoPublicKey.Rsa(
-                        ByteArray((0..10).random()) { 0 } + pubKey.modulus.toByteArray(),
-                        pubKey.publicExponent.toInt()
-                    )
-                val iosDerived: CryptoPublicKey = CryptoPublicKey.fromIosEncoded(own.iosEncoded)
-                val keyIdDerived: CryptoPublicKey = CryptoPublicKey.fromKeyId(own.keyId)
+                val own = CryptoPublicKey.Rsa(pubKey.modulus.toByteArray(), pubKey.publicExponent.toInt())
+                val own1 = CryptoPublicKey.Rsa(
+                    ByteArray((0..10).random()) { 0 } + pubKey.modulus.toByteArray(),
+                    pubKey.publicExponent.toInt()
+                )
 
                 // Correctly drops leading zeros
-                pubKey2.n shouldBe own.n
-                pubKey2.e shouldBe own.e
-
-                pubKey2.hashCode() shouldBe own.hashCode()
-                pubKey2 shouldBe own
-                iosDerived.hashCode() shouldBe own.hashCode()
-                iosDerived shouldBe own
-                keyIdDerived.hashCode() shouldBe own.hashCode()
-                keyIdDerived shouldBe own
+                own1.n shouldBe own.n
+                own1.e shouldBe own.e
 
                 println(Json.encodeToString(own))
                 println(own.iosEncoded.encodeToString(Base16()))
@@ -115,6 +99,13 @@ class PublicKeyTest : FreeSpec({
                 own.getPublicKey().encoded shouldBe pubKey.encoded
             }
         }
+        "Equality tests" {
+            val keyPair = KeyPairGenerator.getInstance("RSA").also { it.initialize(2048) }.genKeyPair()
+            val pubKey1 = CryptoPublicKey.decodeFromDer(keyPair.public.encoded)
+            val pubKey2 = CryptoPublicKey.decodeFromDer(keyPair.public.encoded)
 
+            pubKey1.hashCode() shouldBe pubKey2.hashCode()
+            pubKey1 shouldBe pubKey2
+        }
     }
 })