+++ CHANGELOG.md
@@ -33,3 +33,9 @@
   * Return `KmmResult` for conversions between different key representations ( i.e. `CryptoPublicKey`, `CoseKey` and `JsonWebKey`) 
 
 ### NEXT
+* Dependency Updates
+  * KmmResult 1.5.4
+* Refactor `MultiBaseHelper` to only handle conversion
+* Change `JwsHeader.publicKey` from JsonWebKey to CryptoPublicKey
+* Remove `SignatureValueLength` parameters from JWS- & Cose Algorithm Enum class
+* Remove deprecated functions